Warhub and lobby materials settable by commands
First, make sure you have a block (or air) in you hand.  Then use
/warcfg warhubmaterial:<floor/outline/gate/light> or /zonecfg
lobbymaterial:<floor/outline/gate/light>.

This is from a previous commit but: when you have AIR in your hand and
you change the floor or the floor's outline, that tells War to leave the
user-build or natural original floor in place.

New and cool: setting down the warhub now only clears the blocks needed
for players to find the gates, letting you build a nice warhub that
won't get wiped. Lobbies now also clear only a small path for players
instead of the whole volume.