Prevent players from smuggling items into warzones with ender chests.
This is mainly for servers such as the Warhub, where most players can make
their own warzones. Inexperienced zone makers may place ender chests in
zones which makes their zone susceptible to smuggling items.

This may also apply for zones which would like to use ender chests for
aesthetics.

Hopefully this change shouldn't be a problem. I cannot think of any way
that an ender chest would be used inside of a warzone.