Option to drop inventory on kill
Closes #711, closes #13, closes #11.

Items only drop on kill to prevent players from farming items. Warzone
makers must design their warzone items to fit with this feature if they
choose to enable it. E.g. a warzone might include items that must be
found and then can be stolen on kill.

Thanks to @dylanhansch for testing.